+++ ecoscope/analysis/UD/__init__.py
@@ -1,5 +1,6 @@
-from ecoscope.analysis.UD.etd_range import calculate_etd_range
+from ecoscope.analysis.UD.etd_range import calculate_etd_range, grid_size_from_geographic_extent
 
 __all__ = [
     "calculate_etd_range",
+    "grid_size_from_geographic_extent",
 ]
