+++ ecoscope/base/utils.py
@@ -277,23 +277,3 @@ def hex_to_rgba(input: str) -> tuple:
 def color_tuple_to_css(color: Tuple[int, int, int, int]) -> str:
     # eg [255,0,120,255] converts to 'rgba(255,0,120,1)'
     return f"rgba({color[0]}, {color[1]}, {color[2]}, {color[3]/255})"
-
-
-def grid_size_from_geographic_extent(gdf: gpd.GeoDataFrame, scale_factor: int = 10) -> float:
-    gdf = gdf.to_crs("EPSG:4326")
-
-    def _haversine_diagonal(bbox: BoundingBox):
-        from sklearn.metrics.pairwise import haversine_distances
-        from math import radians
-
-        bbox_radians = [radians(x) for x in bbox]
-        min_point = [bbox_radians[0], bbox_radians[1]]
-        max_point = [bbox_radians[2], bbox_radians[3]]
-        result = haversine_distances([min_point, max_point])
-        result = result * 6371000 / 100  # multiply by Earth radius to get metres
-        return result[0][1]
-
-    local_bounds = tuple(gdf.geometry.total_bounds.tolist())
-    local_cell_size = int(round(_haversine_diagonal(local_bounds) / scale_factor, 0))
-
-    return local_cell_size
